✨ feat: change payment amount type from Long to Double for improved precision

+++ domain/src/main/java/com/mjtech/store/domain/payment/entities/Payment.kt
@@ -4,14 +4,14 @@ import kotlinx.serialization.Serializable
 
 /**
  * @param id O ID do pagamento.
- * @param amount O valor do pagamento, em centavos.
+ * @param amount O valor do pagamento.
  * @param type O tipo de pagamento (crédito, débito, pix, etc.).
  * @param installmentDetails Detalhes do parcelamento, pode ser nulo.
  */
 @Serializable
 data class Payment(
     val id: Long,
-    val amount: Long,
+    val amount: Double,
     val type: PaymentType,
     val installmentDetails: InstallmentDetails?
 )
